auth-api/requirements.txt
fastapi
uvicorn
cachetools
starlette-exporter
boto3
kafka-python
pydantic
requests
mangum